web: reuse community OfficialLinks on confirmation page; lowercase breadcrumbs; don't re-download on refresh
- Extract the Community page's official-links card grid (Discord, GitHub, X,
  YouTube, LinkedIn) into a shared <OfficialLinks> component and reuse it on
  both the Community page and the download confirmation page, so the two stay
  in sync. The confirmation page drops its bespoke Documentation/Discord/X
  cards in favor of these shared components.
- Lowercase the header breadcrumb on both pages ("download", "community") across
  all 20 locales; the page H1 on Community stays "Community".
- Only auto-trigger the download on a fresh forward navigation (e.g. clicking a
  Download button). Refreshing /download/confirmation or using back/forward no
  longer re-downloads; the "Click here" link remains the manual fallback.
