+++ web/app/[locale]/community/page.tsx
@@ -2,6 +2,7 @@ import { useLocale, useTranslations } from "next-intl";
 import { getTranslations } from "next-intl/server";
 import { buildAlternates } from "../../../i18n/seo";
 import { SiteHeader } from "../components/site-header";
+import { OfficialLinks } from "../components/official-links";
 import {
   awesomeCmuxCategoryOrder,
   awesomeCmuxProjectRows,
@@ -18,113 +19,6 @@ export async function generateMetadata({ params }: { params: Promise<{ locale: s
     description: t("metaDescription"),
     alternates: buildAlternates(locale, "/community"),
   };
-}
-
-function CommunityLink({
-  href,
-  icon,
-  name,
-  action,
-  description,
-}: {
-  href: string;
-  icon: React.ReactNode;
-  name: string;
-  action: string;
-  description: string;
-}) {
-  return (
-    <a
-      href={href}
-      target="_blank"
-      rel="noopener noreferrer"
-      className="group flex items-start gap-4 rounded-lg border border-border p-5 hover:bg-code-bg transition-colors"
-    >
-      <div className="shrink-0 mt-0.5 text-muted group-hover:text-foreground transition-colors">
-        {icon}
-      </div>
-      <div className="min-w-0">
-        <div className="font-medium text-[15px]">{name}</div>
-        <div className="text-sm text-muted mt-0.5">{description}</div>
-        <div className="text-xs font-medium text-muted mt-2 group-hover:text-foreground transition-colors">
-          {action} &rarr;
-        </div>
-      </div>
-    </a>
-  );
-}
-
-function OfficialLinks() {
-  const t = useTranslations("community");
-
-  return (
-    <section className="mb-10">
-      <h2 className="mb-4 text-xs font-medium tracking-tight text-muted">
-        {t("officialLinksTitle")}
-      </h2>
-      <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-5">
-        <CommunityLink
-          href="https://discord.gg/xsgFEVrWCZ"
-          name={t("discord")}
-          action={t("discordAction")}
-          description={t("discordDesc")}
-          icon={
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-              <path d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515.074.074 0 0 0-.079.037c-.21.375-.444.864-.608 1.25a18.27 18.27 0 0 0-5.487 0 12.64 12.64 0 0 0-.617-1.25.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057 19.9 19.9 0 0 0 5.993 3.03.078.078 0 0 0 .084-.028c.462-.63.874-1.295 1.226-1.994a.076.076 0 0 0-.041-.106 13.107 13.107 0 0 1-1.872-.892.077.077 0 0 1-.008-.128 10.2 10.2 0 0 0 .372-.292.074.074 0 0 1 .077-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .078.01c.12.098.246.198.373.292a.077.077 0 0 1-.006.127 12.299 12.299 0 0 1-1.873.892.077.077 0 0 0-.041.107c.36.698.772 1.362 1.225 1.993a.076.076 0 0 0 .084.028 19.839 19.839 0 0 0 6.002-3.03.077.077 0 0 0 .032-.054c.5-5.177-.838-9.674-3.549-13.66a.061.061 0 0 0-.031-.03zM8.02 15.33c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.956-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.956 2.418-2.157 2.418zm7.975 0c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.955-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.946 2.418-2.157 2.418z" />
-            </svg>
-          }
-        />
-
-        <CommunityLink
-          href="https://github.com/manaflow-ai/cmux"
-          name="GitHub"
-          action={t("githubAction")}
-          description={t("githubDesc")}
-          icon={
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-              <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z" />
-            </svg>
-          }
-        />
-
-        <CommunityLink
-          href="https://twitter.com/manaflowai"
-          name={t("twitter")}
-          action={t("twitterAction")}
-          description={t("twitterDesc")}
-          icon={
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-              <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z" />
-            </svg>
-          }
-        />
-
-        <CommunityLink
-          href="https://www.youtube.com/channel/UCAa89_j-TWkrXfk9A3CbASw"
-          name={t("youtube")}
-          action={t("youtubeAction")}
-          description={t("youtubeDesc")}
-          icon={
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-              <path d="M23.498 6.186a3.016 3.016 0 0 0-2.122-2.136C19.505 3.545 12 3.545 12 3.545s-7.505 0-9.377.505A3.017 3.017 0 0 0 .502 6.186C0 8.07 0 12 0 12s0 3.93.502 5.814a3.016 3.016 0 0 0 2.122 2.136c1.871.505 9.376.505 9.376.505s7.505 0 9.377-.505a3.015 3.015 0 0 0 2.122-2.136C24 15.93 24 12 24 12s0-3.93-.502-5.814zM9.545 15.568V8.432L15.818 12l-6.273 3.568z" />
-            </svg>
-          }
-        />
-
-        <CommunityLink
-          href="https://www.linkedin.com/company/manaflow-ai/"
-          name={t("linkedin")}
-          action={t("linkedinAction")}
-          description={t("linkedinDesc")}
-          icon={
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-              <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z" />
-            </svg>
-          }
-        />
-      </div>
-    </section>
-  );
 }
 
 const categorySummaries = awesomeCmuxCategoryOrder
@@ -162,7 +56,7 @@ export default function CommunityPage() {
 
   return (
     <div className="min-h-screen">
-      <SiteHeader section={t("title")} />
+      <SiteHeader section={t("section")} />
       <main className="w-full max-w-6xl mx-auto px-6 py-10">
         <h1 className="text-2xl font-semibold tracking-tight mb-2">
           {t("title")}
