+++ web/messages/ar.json
@@ -160,7 +160,8 @@
       "buildDistribution": "البناء والتوزيع",
       "upstreamDependencies": "اعتماديات upstream",
       "archived": "مؤرشف"
-    }
+    },
+    "section": "المجتمع"
   },
   "blog": {
     "title": "المدونة",
